% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_model}
\alias{build_model}
\title{Build models}
\usage{
build_model(.object, ...)
}
\arguments{
\item{.object}{An object of a supported class. See S3 methods below.}

\item{...}{Arguments passed to the appropriate S3 method.}
}
\value{
An object of class \code{tbl_df} (tibble) summarizing the provided
object.
}
\description{
Models specified terms in model data against an existing model
and returns a clean, human readable table of summarizing the effects and
statistics for the newly generated model. This function is meant to simplify
fitting a large number of variables against a set of time-to-event data.
}
\seealso{
\code{\link{build_model.coxph}}
}
