% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_row.R
\name{build_row.data.frame}
\alias{build_row.data.frame}
\title{Summarize a data.frame or tibble}
\usage{
\method{build_row}{data.frame}(
  x,
  y = NA_real_,
  label = "n(\%)",
  stat.pct.sign = FALSE,
  col.overall = TRUE,
  col.missing = FALSE,
  col.test = FALSE,
  digits = 1,
  ...
)
}
\arguments{
\item{x}{An data.frame object. Data to summarize. Must be the same length as
\code{y} (if specified).}

\item{y}{A factor or logical. Data to stratify \code{x} by.}

\item{label}{A character. The name of the summarized variable.}

\item{stat.pct.sign}{A logical. Paste a percentage symbol with each frequency.
frequency.}

\item{col.overall}{A logical. Append a column with the statistic for all data.
If \code{y} is not specified, this parameter is ignored.}

\item{col.missing}{A logical. Append a column with counts of missing data.}

\item{col.test}{A logical. Append a column with the name of the statistical
test used.}

\item{digits}{An integer. Number of digits to round to.}

\item{...}{Additional arguments passed to S3 method.}
}
\value{
An object of class \code{tbl_df} (tibble) summarizing the provided
data.
}
\description{
Summarize a data.frame (row counts). Optional stratification
using a factor or logical with the same size as the tibble.
}
\examples{
# Create a "count" row from a data.frame for a factor
build_row(x = datasets::mtcars, y = as.factor(datasets::mtcars$cyl))
}
