% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_row.R
\name{build_row.numeric}
\alias{build_row.numeric}
\title{Summarize numeric data}
\usage{
\method{build_row}{numeric}(
  x,
  y = NA_real_,
  label = "< variable >",
  label.stat = TRUE,
  stat = c("mean", "median"),
  stat.pct.sign = FALSE,
  col.overall = TRUE,
  col.missing = FALSE,
  test = c("anova", "kruskal", "wilcoxon"),
  col.test = FALSE,
  digits = 1,
  p.digits = 4,
  ...
)
}
\arguments{
\item{x}{A numeric. Data to summarize. Must be the same length as \code{y}
(if specified).}

\item{y}{A factor or logical. Data to stratify \code{x} by.}

\item{label}{A character. The name of the summarized variable.}

\item{label.stat}{A logical. Append the summary statistic used to
the label of the summarized row.}

\item{stat}{A character. Name of the summary statistic to use. Supported options
include the mean ('mean') and median ('median').}

\item{stat.pct.sign}{A logical. Paste a percentage symbol with each frequency.}

\item{col.overall}{A logical. Append a column with the statistic for all data.
If \code{y} is not specified, this parameter is ignored.}

\item{col.missing}{A logical. Append a column with counts of missing data.}

\item{test}{A character. Name of statistical test to compare groups.
Supported options include ANOVA linear model ('anova'), Kruskal-Wallis ('kruskal'),
and Wilcoxon rank sum ('wilcoxon') tests.}

\item{col.test}{A logical. Append a column with the name of the statistical
test used.}

\item{digits}{An integer. Number of digits to round to.}

\item{p.digits}{An integer. Number of p-value digits to report.}

\item{...}{Additional arguments passed to S3 method.}
}
\value{
An object of class \code{tbl_df} (tibble) summarizing the provided
data.
}
\description{
Summarize numeric data in a tibble. Optional stratification and
null hypothesis testing using another factor or logical.
}
\examples{
# Create a row summarizing a numeric by a factor
build_row(
 x = datasets::mtcars$mpg,
 y = as.factor(datasets::mtcars$cyl),
 label = 'MPG'
)
}
\seealso{
\code{\link{build_row}}
}
