% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_row.R
\name{build_row}
\alias{build_row}
\title{Build summary rows}
\usage{
build_row(x, ...)
}
\arguments{
\item{x}{An object of type data.frame, logical, factor, or numeric.}

\item{...}{S3 method specific arguments.}
}
\description{
Build summary rows
}
\examples{
library(dplyr)

data_mtcars <- datasets::mtcars \%>\%
  as_tibble() \%>\%
  mutate_at(vars('vs', 'am'), as.logical) \%>\%
  mutate_at(vars('gear', 'carb', 'cyl'), as.factor)

# Create a "count" row from a data.frame
build_row(x = data_mtcars, y = data_mtcars$cyl)

# Create a row summarizing a numeric by a factor
build_row(x = data_mtcars$mpg, y = data_mtcars$cyl)

# Create a row summarizing a factor by a factor
build_row(x = data_mtcars$carb, y = data_mtcars$cyl)

# Create a row summarizing a logical by a factor
build_row(x = data_mtcars$vs, y = data_mtcars$cyl)
}
