% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_table.R
\name{build_table.data.frame}
\alias{build_table.data.frame}
\title{Build summary tables from data.frame objects}
\usage{
\method{build_table}{data.frame}(
  .object,
  ...,
  .by,
  .inverse = FALSE,
  .append.stat = TRUE,
  .parametric = FALSE,
  .show.missing = FALSE,
  .show.test = FALSE,
  .na.rm = TRUE,
  .percent.sign = TRUE,
  .digits = 1,
  .p.digits = 4
)
}
\arguments{
\item{.object}{A data.frame.}

\item{.by}{An unquoted expression representing a column to stratify summaries
by.}

\item{.inverse}{A logical. For logical data, report the frequency of FALSE
values instead of the TRUE.}

\item{.append.stat}{A logical. Append the summary statistic to the column
label.}

\item{.parametric}{A logical. Use parametric tests as opposed to
non-parametric.}

\item{.show.missing}{A logical. Append a column listing the frequencies of
missing data for each row.}

\item{.show.test}{A logical. Append a column containing the test each p-value
was derived from.}

\item{.na.rm}{A logical. Ignore NA values when calculating the frequencies
for logical and factor data types.}

\item{.percent.sign}{A logical. Paste a percent symbol after all reported
frequencies.}

\item{.digits}{An integer. The number of digits to round numbers to.}

\item{.p.digits}{An integer. The number of p-value digits to report. Note
that the p-value still rounded to the number of digits specified in
\code{.digits}.}
}
\description{
Build summary tables from data.frame objects
}
