% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_table.R
\name{build_table}
\alias{build_table}
\title{Build summary tables}
\usage{
build_table(.object, ...)
}
\arguments{
\item{.object}{An object of a supported class. See S3 methods below}

\item{...}{Arguments passed to the appropriate S3 method.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} summarizing the
provided object.
}
\description{
Build summary tables
}
\seealso{
\code{\link{build_table.data.frame}}
}
