% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_row.R
\name{build_row.numeric}
\alias{build_row.numeric}
\title{Summarize numeric data}
\usage{
\method{build_row}{numeric}(
  x,
  y = NA,
  label = "(Unlabeled column)",
  parametric = FALSE,
  append.stat = TRUE,
  show.missing = FALSE,
  show.test = FALSE,
  percent.sign = TRUE,
  digits = 1,
  p.digits = 4,
  ...
)
}
\arguments{
\item{x}{A numeric. Data to summarize. Must be the same length as \code{y}
(if specified).}

\item{y}{A factor or logical. Optional. Data to stratify \code{x} by.}

\item{label}{A character. Optional. The name of the summarized variable.}

\item{parametric}{A logical. Optional. Use parametric tests.}

\item{append.stat}{A logical. Optional. Append the summary statistic used to
the label of the summarized row.}

\item{show.missing}{A logical. Optional. Append summary counts of missing
data.}

\item{show.test}{A logical. Optional. Show the statistical test and test
statistic used to determine the p-value.}

\item{percent.sign}{A logical. Optional. Paste a percentage symbol with each
frequency.}

\item{digits}{An integer. Optional. Number of digits to round to.}

\item{p.digits}{An integer. Optional. Number of p-value digits to report.}

\item{...}{Miscellaneous options.}
}
\value{
An object of class \code{tbl_df} (tibble) summarizing the provided
data.
}
\description{
Summarize numeric data in a tibble. Optional stratification and
null hypothesis testing using another factor or logical.
}
\examples{
# Create a row summarizing a numeric by a factor
build_row(
 x = datasets::mtcars$mpg,
 y = as.factor(datasets::mtcars$cyl),
 label = 'MPG'
)
}
\seealso{
\code{\link{build_row}}
}
