% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_row.R
\name{build_row}
\alias{build_row}
\title{Build summary rows}
\usage{
build_row(x, ...)
}
\arguments{
\item{x}{An object of a supported class. See S3 methods below.}

\item{...}{Arguments passed to the appropriate S3 method.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} summarizing the
provided data.
}
\description{
Summarize a data into a data.frame row(s). Optional
stratification and null hypothesis testing using a factor or logical.
}
\seealso{
\code{\link{build_row.data.frame}},
\code{\link{build_row.numeric}},
\code{\link{build_row.logical}},
\code{\link{build_row.factor}}
}
