% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_table.R
\name{build_table}
\alias{build_table}
\title{Build summary tables}
\usage{
build_table(.object, ...)
}
\arguments{
\item{.object}{An object of a supported class. See S3 methods below}

\item{...}{Arguments passed to the appropriate S3 method.}
}
\value{

}
\description{
Build summary tables
}
\examples{
library(dplyr)

data_mtcars <- datasets::mtcars \%>\%
  as_tibble() \%>\%
  mutate_at(vars('vs', 'am'), as.logical) \%>\%
  mutate_at(vars('gear', 'carb', 'cyl'), as.factor)

# Summarize all columns by cylindars variable
data_mtcars \%>\% build_table(by = cyl, show.test = TRUE)

# Summarize specific columns of data
data_mtcars \%>\% build_table(mpg, vs, carb)

# Summarize columns using tidyselect helpers
data_mtcars \%>\% build_table(starts_with('c'), mpg, by = am)
}
\seealso{
\code{\link{build_table.data.frame}}
}
