% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_table.R
\name{build_table}
\alias{build_table}
\title{Build summary tables}
\usage{
build_table(.object, ...)
}
\arguments{
\item{.object}{An object of a supported class. See S3 methods below.}

\item{...}{Arguments passed to the appropriate S3 method.}
}
\value{
An object of class tbl_df (tibble) summarizing the provided object.
}
\description{
Takes a data or model object and summarizes it into a ready to
export, human-readable summary table.
}
\seealso{
\code{\link{build_table.data.frame}},
\code{\link{build_table.coxph}},
\code{\link{build_table.lm}}
}
